Update ibeo_core.h
PointClouds with more than ~2323 Points will be parsed wrong!
Fix offset type to enable lager pcl

24 + (3*148) + (2324 * 28) = 65.540
24 + (3*148) + (2323 * 28) = 65.512
2^16 = 65.536

+++ include/ibeo_core/ibeo_core.h
@@ -343,7 +343,7 @@ public:
 
   void parse(
     const std::vector<uint8_t>& in,
-    const uint16_t& offset);
+    const uint32_t& offset);
 };
 
 class ScanPoint2204
@@ -363,7 +363,7 @@ public:
 
   void parse(
     const std::vector<uint8_t>& in,
-    const uint16_t& offset);
+    const uint32_t& offset);
 };
 
 class ScanPoint2205
@@ -384,7 +384,7 @@ public:
 
   void parse(
     const std::vector<uint8_t>& in,
-    const uint16_t& offset);
+    const uint32_t& offset);
 };
 
 class ScanPoint2208
@@ -402,7 +402,7 @@ public:
 
   void parse(
     const std::vector<uint8_t>& in,
-    const uint16_t& offset);
+    const uint32_t& offset);
 };
 
 struct IbeoObject
